Move ViewModel to shared module to prepare for KMP
