+++ Fruitties/shared/src/commonMain/kotlin/com/example/fruitties/viewmodel/MainViewModel.kt
@@ -14,16 +14,12 @@
  * limitations under the License.
  */
 
-package com.example.fruitties.android.ui
+package com.example.fruitties.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.ViewModelProvider.AndroidViewModelFactory.Companion.APPLICATION_KEY
 import androidx.lifecycle.viewModelScope
-import androidx.lifecycle.viewmodel.initializer
-import androidx.lifecycle.viewmodel.viewModelFactory
 import com.example.fruitties.DataRepository
-import com.example.fruitties.android.di.App
 import com.example.fruitties.database.CartItemDetails
 import com.example.fruitties.model.Fruittie
 import kotlinx.coroutines.flow.SharingStarted
@@ -57,15 +53,11 @@ class MainViewModel(private val repository: DataRepository) : ViewModel() {
     }
 
     companion object {
-        val Factory: ViewModelProvider.Factory = viewModelFactory {
-            initializer {
-                val application = (this[APPLICATION_KEY] as App)
-                val repository = application.container.dataRepository
-                MainViewModel(repository = repository)
-            }
-        }
+        val Factory: ViewModelProvider.Factory = createMainViewModelFactory()
     }
 }
+
+expect fun createMainViewModelFactory(): ViewModelProvider.Factory
 
 /**
  * Ui State for ListScreen
